+++ examples/eks_test_fixture/main.tf
@@ -11,6 +11,21 @@ provider "random" {
   version = "= 1.3.1"
 }
 
+data "aws_eks_cluster" "cluster" {
+  name = "${module.eks.cluster_id}"
+}
+
+data "aws_eks_cluster_auth" "cluster" {
+  name = "${module.eks.cluster_id}"
+}
+
+provider "kubernetes" {
+  host                   = "${data.aws_eks_cluster.cluster.endpoint}"
+  cluster_ca_certificate = "${base64decode(data.aws_eks_cluster.cluster.certificate_authority.0.data)}"
+  token                  = "${data.aws_eks_cluster_auth.cluster.token}"
+  load_config_file       = false
+}
+
 data "aws_availability_zones" "available" {}
 
 locals {