+++ CHANGELOG.md
@@ -18,6 +18,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
  - Add .prettierignore file (by @rothandrew)
  - Switch to https for the pre-commit repos (by @rothandrew)
  - Add instructions on how to enable the docker bridge network (by @rothandrew)
+ - Changed aws auth to be managed using kubernetes provider (by @stijndehaes)
  - Write your awesome change here (by @you)
 
 # History